+++ source/iked/ike.h
@@ -559,9 +559,9 @@ typedef struct _IKE_PEER
 	unsigned short	natt_port;
 	long			natt_rate;
 	long			frag_ike_mode;
-	long			frag_ike_size;
+	size_t			frag_ike_size;
 	long			frag_esp_mode;
-	long			frag_esp_size;
+	size_t			frag_esp_size;
 	long			dpd_mode;
 	long			dpd_rate;
 	long			life_check;
